# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Social Good Web3 system
  description: Social Good API
  version: 1.0.0
servers:
  - url: http://localhost:50001
    description: API Mock
paths:
  /v1/wallet/{wallet_address}/user_id:
    get:
      security:
        - Bearer: []
      tags:
        - Wallet
      summary: Get User ID by wallet_address
      operationId: getUserIdByWalletAddress
      parameters:
        - name: jwt
          in: header
          description: IdToken
          required: true
          schema:
            type: string
          example: Bearer eyxxx
        - name: public_key
          in: header
          description: Public Key
          required: true
          schema:
            type: string
          example: 02d9fbeeb08f28182c10733f79180bcace5ef68b2841c83252e3574be1d97492e2
        - name: wallet_address
          in: path
          description: Wallet Address
          required: true
          schema:
            type: string
          example: '0x60f80121c31a0d46b5279700f9df786054aa5ee6'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUnauthorized'
        '500':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseInternalServerError'
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Credentials or access token for API
  schemas:
...
    ResponseInternalServerError:
      type: object
      properties:
        message:
          type: string
          example: Service Unavailable
...
    ResponseUnauthorized:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
...
    WalletUser:
      type: object
      properties:
        user_id:
          type: string
          example: 9490ebf5-e378-4074-97d0-3ce5004ae918